pyproject.toml
[project]
name = "gridpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "fastapi>=0.135.1",
    "markdown>=3.10.2",
    "pyserial>=3.5",
]
